requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
chardet==5.2.0
click==8.1.7
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
instructor==1.3.5
jiter==0.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
numpy==2.0.0
ollama==0.3.0
openai==1.59.9
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyreadstat==1.2.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
tenacity==8.5.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
yarl==1.9.4
